+++ src/googleTTS2/GoogleTTS2.ts
@@ -12,6 +12,7 @@ import fs from 'fs';
 import { CommonTTS } from '../common/commonTTS';
 import { GoogleTTS2Request } from './GoogleTTS2Request';
 import { GoogleVoices } from './GoogleTTS2Voices';
+import pc from 'picocolors';
 
 export interface GoogleToken {
     "type": "service_account",
@@ -56,14 +57,17 @@ export class GoogleTTS2 extends CommonTTS<GoogleTTS2Request> {
         const {file, exists} = await this.cacheDir.getCacheFile(request.hash(), request.filename());
         if (exists)
             return file;
-        const tokenFile = await this.getToken();
+        // ensure GOOGLE_APPLICATION_CREDENTIALS env is set
+        await this.getToken();
 
         try {
+            const start = Date.now();
             const [responce, error] = await this.client.synthesizeSpeech(request.toRequest());
             let data = responce.audioContent;
             if (!data) {
-                throw Error('Access VoiceMaker failed with response ' + JSON.stringify(error));
+                throw Error(`Access VoiceMaker failed with response ${JSON.stringify(error)}`);
             }
+            console.log(`new speech generated in ${pc.yellow(Date.now() - start)}ms`);
             await fs.promises.writeFile(file, data);
             await super.log(request);
         } catch (e) {